+++ src/shapes/Text.ts
@@ -95,9 +95,11 @@ function checkDefaultFill(config) {
 }
 
 // polyfill for IE11
-const trimRight = String.prototype.trimRight || function polyfill() {
-  return this.replace(/[\s\xa0]+$/, '');
-}
+const trimRight =
+  String.prototype.trimRight ||
+  function polyfill() {
+    return this.replace(/[\s\xa0]+$/, '');
+  };
 
 /**
  * Text constructor
@@ -294,7 +296,11 @@ export class Text extends Shape<TextConfig> {
     context.fillStrokeShape(this);
   }
   setText(text) {
-    var str = Util._isString(text) ? text : (text === null || text === undefined) ? '' : text + '';
+    var str = Util._isString(text)
+      ? text
+      : text === null || text === undefined
+      ? ''
+      : text + '';
     this._setAttr(TEXT, str);
     return this;
   }